refactorización de paquetes
cambios en la estructura de paquetes para tener una estructura común y con nombres más significativos (y sin ñ's)
