+++ src/main/java/cv/gui/dialogs/ConocimientoDialog.java
@@ -1,11 +1,11 @@
-package gui.Entrañas.Dialog;
+package cv.gui.dialogs;
 
 import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-import cv.Conocimiento;
-import cv.Nivel;
+import cv.model.Conocimiento;
+import cv.model.Nivel;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
