+++ src/main/java/cv/model/Email.java
@@ -1,4 +1,4 @@
-package cv;
+package cv.model;
 
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
