+++ src/main/java/cv/model/Experiencia.java
@@ -1,4 +1,4 @@
-package cv;
+package cv.model;
 
 import java.time.LocalDate;
