+++ src/main/java/cv/model/Personal.java
@@ -1,4 +1,4 @@
-package cv;
+package cv.model;
 
 import java.time.LocalDate;
